+++ client/src/scene-components/GroundComponent.ts
@@ -1,8 +1,9 @@
 import {
+  GridHelper,
+  Group,
   Mesh,
-  MeshStandardMaterial,
+  MeshPhongMaterial,
   PlaneGeometry,
-  type Texture,
   Vector3,
   type Scene
 } from "three";
@@ -12,35 +13,44 @@ import type { SceneComponent } from "../types/SceneComponent";
 type GroundState = "idle";
 
 export class GroundComponent implements SceneComponent<GroundState> {
+  private readonly root: Group;
   private readonly ground: Mesh;
+  private readonly grid: GridHelper;
   private readonly position = new Vector3();
 
-  public constructor(groundTexture: Texture) {
+  public constructor() {
     this.ground = new Mesh(
-      new PlaneGeometry(600, 600),
-      new MeshStandardMaterial({ color: 0x8c9b75 })
+      new PlaneGeometry(2000, 2000),
+      new MeshPhongMaterial({
+        color: 0x999999,
+        depthWrite: false
+      })
     );
     this.ground.rotation.x = -Math.PI / 2;
     this.ground.receiveShadow = true;
 
-    const material = this.ground.material as MeshStandardMaterial;
-    material.map = groundTexture;
-    material.needsUpdate = true;
+    this.grid = new GridHelper(2000, 40, 0x000000, 0x000000);
+    this.grid.material.opacity = 0.2;
+    this.grid.material.transparent = true;
+
+    this.root = new Group();
+    this.root.add(this.ground);
+    this.root.add(this.grid);
   }
 
   public add(scene: Scene): void {
-    scene.add(this.ground);
+    scene.add(this.root);
   }
 
   public remove(scene: Scene): void {
-    scene.remove(this.ground);
+    scene.remove(this.root);
   }
 
   public tick(_deltaTime: number): void {}
 
   public setPosition(x: number, y: number, z: number): void {
     this.position.set(x, y, z);
-    this.ground.position.copy(this.position);
+    this.root.position.copy(this.position);
   }
 
   public transition(state: GroundState): void {